libsbom/spdx: Correctly access build metadata properties
When adding annotations for build properties to an SBOM, the SPDX
SBOM generator iterates over the metadata dictionary instead of the
property dictionary (which is contained in the former), causing an
exception to be thrown.

This commit changes the `SBOM.add_build()` method so that it
iterates over the "properties" member of the metadata dictionary
instead of the metadata dictionary.