libsbom/spdx: Use timestamp from build when component doesn't have one
The SPDX SBOM generator expects the build timestamp of a package to be
contained in the component metadata, no matter what kind of input data
it is handling. However, if the input data contains multiple components,
the build timestamp is in the build metadata, not in the components.
Thus, the SPDX generator attempts to parse invalid data, causing an
exception to be thrown.

This commit modifies the SPDX generator so that it attempts to read the
timestamp from the build metadata if it could not be parsed from the
component metadata.