libsbom: Normalize epochs in package versions
When a package's epoch is not set, alma-sbom generates SBOMs with
the package's version set to something like "None:1.0". Since the
epoch has to be an unsigned integer (if it is present at all),
this might confuse tools that are consuming SBOM data.
According to the RPM spec, a missing epoch is equivalent to 0, so
this is what should be in the SBOM.

This commit modifies the SPDX and CycloneDX SBOM generators so
that missing epochs will be normalized to 0.