libsbom: Use same vendor name in both SBOM formats
When generating SBOMs in CycloneDX format, alma-sbom uses a hardcoded
vendor name instead of the name in the constants module. Thus, SBOMs
in CycloneDX format have a different vendor than those in SPDX format.

This commit modifies the CycloneDX SBOM generation so it uses the
vendor name from the constants module instead of a hardcoded value.