Refactor input_data structure.
The information generated by get_info_about_build() and get_info_about_package()
is structured similarly to the CycloneDX data model.
However, the way information is used in the metadata field was different from
that of CycloneDX.
This patch fixed that.
This change will be useful when adding information to the metadata field in the future.

In addition, I have slightly changed the way SBOMs are created in libsbom/spdx.py
to use the sbom_object_type.