Fix some incomplete metadata issue partially
When data cannot be retrieved from immudb, or when data for package is
missing, if the --rpm-package option was used to specify an rpm package,
I made it so that the rpm package is used to supplement the information.

The following issues have been partially fixed.
 - https://github.com/AlmaLinux/alma-sbom/issues/42
 - https://github.com/AlmaLinux/alma-sbom/issues/44

And the following issues have been fixed.
 - https://github.com/AlmaLinux/alma-sbom/issues/26